ai-hackathon-use-cases/Constituent-Services-Agent/requirements.txt
# Core Azure AI
azure-ai-projects>=1.0.0
azure-ai-evaluation>=1.13.7
azure-ai-inference>=1.0.0
azure-identity>=1.19.0

# Microsoft Agent Framework
semantic-kernel>=1.45.0

# Azure AI Search
azure-search-documents>=11.5.0

# Translation for multi-language support
azure-ai-translation-text>=1.0.0

# Text Analytics
azure-ai-textanalytics>=5.3.0

# OpenAI
openai>=1.98.0

# Web Framework
flask>=3.1.0

# Utilities
python-dotenv>=1.0.1
requests>=2.32.0
aiohttp>=3.11.0

# Testing
pytest>=8.0.0
pytest-asyncio>=0.24.0

# Evaluation tools
promptfoo>=0.100.0
